checkpoint 9 working

+++ app/controllers/likes_controller.rb
@@ -2,7 +2,7 @@ class LikesController < ApplicationController
   def create
     @bookmark = Bookmark.find(params[:bookmark_id])
     like = current_user.likes.build(bookmark: @bookmark)
-    # authorize @like
+    authorize like
 
     if like.save
       flash[:notice] = 'Liked!'
@@ -16,7 +16,7 @@ class LikesController < ApplicationController
   def destroy
     @bookmark = Bookmark.find(params[:bookmark_id])
     like = current_user.likes.find(params[:id])
-    # authorize @like
+    authorize like
 
     if like.destroy
       flash[:notice] = 'Unliked!'
